\name{multgam-package}
\alias{multgam-package}
\alias{multgam}
\docType{package}
\title{
\packageTitle{multgam}
}
\description{
\packageDescription{multgam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{multgam}
\packageIndices{multgam}
~~ see https://github.com/yelbachir/multgam/ ~~
}
\author{
\packageAuthor{multgam}

Maintainer: \packageMaintainer{multgam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
